Sort of.
This is not exactly how you want things to be. I’m toggling the
classes, so it will work fine as long as you untoggle before clicking
the next filter you desire. Need to add conditional statements instead
of toggle so that the links can turn highlight on/off and are not
dependent on one another.